[chore] Move pcf module
