+++ src/gmx2qmmm/generators/__init__.py
@@ -1,0 +1,4 @@
+from gmx2qmmm.generators.pcf.pcf import PCFGeneratorShift
+
+
+__all__ = ["PCFGeneratorShift"]
