+++ src/gmx2qmmm/jobs/qm.py
@@ -21,7 +21,7 @@ import numpy as np
 #   Imports From Custom Libraries
 from gmx2qmmm.generators._helper import filter_xyzq, _flatten
 from gmx2qmmm.generators.geometry import read_gmx_structure_header, read_gmx_structure_atoms, read_gmx_box_vectors, write_g96
-from gmx2qmmm.generators.pcf import PCFGenerator
+from gmx2qmmm.generators.pcf.base import PCFGenerator
 
 #   // TODOS & NOTES //
 #   TODO:
